// cli/src/core/api/types.ts
export interface PaginatedResponse<T> {
  totalPages: number;
  totalElements: number;
  data: T[];
}

export interface PaginationParams {
  limit?: number;
  offset?: number;
}
// ...
export interface TaxProfile {
  resourceId: string;
  name: string;
  taxRate: number;
  taxTypeCode: string;
  status: string;
}

export interface TaxType {
  resourceId: string;
  code: string;
  name: string;
}
// ...
export interface WithholdingTaxCode {
  resourceId: string;
  code: string;
  countryCode: string;
  description: string;
  standard: string;
}